Collision fixes.
git-svn-id: file:///home/cbm/dev/rpgtoolkit/svn/trunk@1636 8ad37bdc-5837-4682-a52a-fbb6d9f1e9f7

+++ vc/trans3/movement/movement.h
@@ -19,7 +19,7 @@
  * Definitions.
  */
 const int MILLISECONDS	= 1000;			// Milliseconds in a second.
-const double PX_FACTOR	= 4.0;			// Movement scaler factor.
+const int PX_FACTOR		= 4;			// Movement scaler factor.
 										// Note: Possibly out by a factor of 2.
 
 // m_pos.loopFrame idle states. Only condition: must be negative.
